% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/src-vertica.R
\name{db_drop_view}
\alias{db_drop_view}
\title{Like db_drop_table, but for views.}
\usage{
db_drop_view(con, view)
}
\arguments{
\item{con}{dplyr src connection object}

\item{view}{The name of the view in DB to drop.}
}
\description{
Like db_drop_table, but for views.
}

