% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/src-vertica.R
\name{db_load_from_file}
\alias{db_load_from_file}
\title{Loads a file (typically CSV) from disk to Vertica.}
\usage{
db_load_from_file(dest, table.name, file.name, sep = " ", skip = 1L,
  append = FALSE)
}
\arguments{
\item{dest}{Vertica connection src to the DB.}

\item{table.name}{The name of the table created in Vertica matching the schema
of the file.}

\item{file.name}{Path to file with data.}

\item{sep}{Delimiter in the file.}

\item{skip}{Number of lines to skip at the beginning of the file (useful if contains headers).}

\item{append}{TRUE if copied data will be added to existing data in table. FALSE to overwrite table data.}
}
\value{
A new tbl_vertica reference to the loaded table.
}
\description{
Currently, this will only work for tables that already exist with compatible
schema for the table.
}
\details{
This loading is fast and is used (by default) in copy_to.
}
\examples{
\dontrun{
vertica <- src_vertica("VerticaDSN")
foo <- db_load_from_file(vertica,"foo","./foo.csv",sep=",")
}
}

