% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{list_udf}
\alias{list_udf}
\title{This function shows to the user the names of the funtions, their type, as well as the return and input data types.}
\usage{
list_udf(src, type = NULL)
}
\arguments{
\item{src}{The src_vertica instance from which to query for UDxes.}

\item{type}{The UDF type to list (one of "aggregate","scalar", or "transform").}
}
\value{
A data frame containing four columns: function name, function type, output data type(s), and input data type(s).
}
\description{
This function shows to the user the names of the funtions, their type, as well as the return and input data types.
}
\examples{
\dontrun{
vertica <- src_vertica("VerticaDSN")
UDF <- list_udf(vertica)
}
}

