% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vertica-functions.R
\name{list_udx}
\alias{list_udx}
\title{This function shows to the user the names of the funtions, their type, as well as the return and input data types.}
\usage{
list_udx(src)
}
\arguments{
\item{src}{The src_vertica instance from which to query for UDxes.}
}
\value{
A data frame containing four columns: function name, function type, output data type(s), and input data type(s).
}
\description{
This function shows to the user the names of the funtions, their type, as well as the return and input data types.
}
\examples{
\dontrun{
vertica <- src_vertica("VerticaDSN")
UDxes <- list_udx(vertica)
}
}

