% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/src-vertica.R
\name{mutate}
\alias{mutate}
\title{Like dplyr::mutate, but allows for pre-evaluation to get column names (useful for multi-column returning functions)}
\usage{
mutate(.data, ..., evalNames = FALSE, collapse = TRUE)
}
\arguments{
\item{.data}{dplyr tbl}

\item{...}{table columns (i.e., as used in dplyr::mutate())}

\item{evalNames}{pre-evaluate the statement to get column names of result (useful for functions that return more than one column)}

\item{collapse}{collapse query into subquery (DEFAULT is TRUE); FALSE is needed for certain procedures quiring that the procedure name cannot be a part of a subquery}
}
\value{
a tbl_vertica object
}
\description{
Like dplyr::mutate, but allows for pre-evaluation to get column names (useful for multi-column returning functions)
}

