% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/src-vertica.R
\name{src_vertica}
\alias{src_vertica}
\title{Establishes a connection to Vertica and returns a dplyr 'src' object, src_vertica, using either vRODBC or JDBC.}
\usage{
src_vertica(dsn = NULL, jdbcpath = NULL, dbname = NULL, host = NULL,
  port = 5433, user = NULL, password = "", load_udf = TRUE)
}
\arguments{
\item{dsn}{The name of the ODBC DSN to use for connecting to Vertica. Must be a string. If NULL, JDBC-use is assumed.}

\item{jdbcpath}{(Only required for JDBC) Path to the JDBC driver on your system.}

\item{dbname}{(Only required for JDBC) The name of the Vertica database instance that you plan to connect to.}

\item{host}{(Only required for JDBC) Host IP of the machine hosting the Vertica instance.}

\item{port}{(Only required for JDBC) Port number of the Vertica server. The Vertica default is 5433.}

\item{user}{(Only required for JDBC) Username of the owner of the database.}

\item{password}{(Only required for JDBC) Password, if any, of the database.}

\item{load_udf}{A boolean value. If TRUE (default), this will automatically register the UDFs into the session.}
}
\value{
A dplyr::src object, src_vertica, to be used with dplyr functions.
}
\description{
Provide either the DSN to use ODBC, or set DSN to NULL to use JDBC.
}
\examples{
\dontrun{
vertica_connection <- src_vertica(dsn="VerticaDSN")
vertica_connection <- src_vertica(dsn=NULL,jdbcpath="/opt/vertica/java/lib/vertica_jdbc.jar",
                                  dbname="foo",host="localhost",port=5433,
                                  user="dbadmin",password="secret")
}
}

