% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vertica.dplyr.R
\docType{package}
\name{vertica.dplyr}
\alias{vertica.dplyr}
\alias{vertica.dplyr-package}
\title{Vertica.dplyR is an R package developed by HP that provides Vertica-backend support for the [dplyr][1] package. Besides support of standard dplyr functionality (such as table manipulation), vertica.dplyr also features:}
\description{
* [HPdata][4]-style functions (data transport from Vertica to Distributed R via the data loader), namely tbl2d**object** (where **object** is either **array** or **frame**), that are compatible with dplyr tbl objects.
* Easy CSV-loading into Vertica.
* copy_to functionality that takes advantage of Vertica's fast **COPY LOCAL** feature.
* Connectivity to Vertica through either JDBC or ODBC.
* (Future Feature) Seamless Vertica UDx invocation.
}
\details{
For more information, please read the PDF or HTML of the full user guide in vignettes/.
}

