% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/src-vertica.R
\name{select}
\alias{select}
\title{Like dplyr::select, but allows for the first argument to be a src_vertica object
for SELECT statements without FROM clauses.}
\usage{
select(.arg, ...)
}
\arguments{
\item{.arg}{dplyr tbl OR src_vertica connection object}

\item{...}{table columns (i.e., as used in mutate())}
}
\value{
a tbl_vertica object
}
\description{
Like dplyr::select, but allows for the first argument to be a src_vertica object
for SELECT statements without FROM clauses.
}
\examples{
\dontrun{
vertica <- src_vertica("VerticaDSN")
table <- select(vertica,foo=some_fun())
table2 <- select(table,some_col_in_table)
}
}

